frontend/requirements.txt
streamlit==1.40.1
streamlit_folium==0.23.0
pandas==2.2.3
python-dotenv==1.0.1
geopandas==1.0.1
folium-vectorgrid==0.1.3
httpx==0.28.1
